Handles two edge cases in genbank parsing, and fixes a GUID bug.
First edge case is identical transcripts or features under the same collection. The duplicates are dropped, and a warning is raised.

The second edge case is CDS intervals that exceed the bounds of their associated Transcript. This is sometimes produced by mangled files that come out of Geneious. This is now a warning, and not an exception.

Discovered that Strand was not being taken into account when generating Interval GUIDs, which led to identical intervals on opposite strands being incorrectly flagged as a duplicate.